+++ firmware/drivers/isis_antenna/isis_antenna_delay.c
@@ -25,7 +25,7 @@
  * 
  * \author Gabriel Mariano Marcelino <gabriel.mm8@gmail.com>
  * 
- * \version 0.7.45
+ * \version 1.0.0
  * 
  * \date 2016/09/23
  * 
@@ -43,7 +43,7 @@
 
 void isis_antenna_delay_s(uint8_t s)
 {
-    vTaskDelay(pdMS_TO_TICKS(1000U*s));
+    vTaskDelay(pdMS_TO_TICKS(1000U*(uint32_t)s));
 }
 
 void isis_antenna_delay_ms(uint16_t ms)
